render_project/requirements.txt
requests>=2.28.0
tqdm>=4.64.0
python-dotenv>=0.20.0
huggingface-hub>=0.13.0
Pillow>=9.0.0
numpy>=1.22.0
pydantic>=1.10.0